fix(check): move type-only import regression test
