+++ crates/vize_canon/src/tests.rs
@@ -1,5 +1,7 @@
 //! Snapshot tests for vize_canon.
 
+mod type_only_import_anchors;
+
 #[cfg(test)]
 #[allow(clippy::disallowed_macros)]
 mod virtual_ts_tests {
@@ -57,35 +59,6 @@ const props = defineProps<Props>()
 
         let virtual_ts = generate_virtual_ts_from_sfc(source);
         insta::assert_snapshot!("virtual_ts_with_props", virtual_ts);
-    }
-
-    #[test]
-    fn virtual_ts_does_not_value_anchor_type_position_imports() {
-        let source = r#"<script setup lang="ts">
-import { GetStudyHistoriesQuery } from './schema'
-import { UnwrapArray } from './types'
-
-type StudyHistory = UnwrapArray<
-  GetStudyHistoriesQuery['student']['studyHistories']
->
-
-defineProps<{ studyHistory: StudyHistory }>()
-</script>
-
-<template>
-  <article>{{ studyHistory }}</article>
-</template>"#;
-
-        let virtual_ts = generate_virtual_ts_from_sfc(source);
-
-        assert!(
-            !virtual_ts.contains("void GetStudyHistoriesQuery"),
-            "type-only import must not be anchored as a value:\n{virtual_ts}"
-        );
-        assert!(
-            !virtual_ts.contains("void UnwrapArray"),
-            "type helper import must not be anchored as a value:\n{virtual_ts}"
-        );
     }
 
     #[test]
